feat(OC*): Separate out tests
