+++ .operator-shared/index/opencode-v2.md
@@ -21,15 +21,20 @@ read_if: Working in packages/opencode-v2 or changing OpenCode V2 preamble inject
 ### `packages/opencode-v2/src/`
 
 - `index.ts` — V2 plugin definition, context-hook preamble injection, command wiring, and detached Helper update trigger
-- `index.test.ts` — Plugin setup, command registration, update trigger, and context injection coverage
 - `commands.ts` — V2 command transforms, Helper execution, XML output framing, and prompt admission
-- `commands.test.ts` — User-command preservation, ordered Helper execution, output framing, and prompt admission coverage
 - `notifications.ts` — Typed server-to-TUI status method and notification event contract
 - `preamble.ts` — Core-backed immutable per-session preamble loading and caching
-- `preamble.test.ts` — Session immutability coverage
-- `tui.tsx`, `tui.test.ts` — RPC status client, connecting/readiness and partition UI, idle refresh, notification listener, and focused coverage
-- `update.ts`, `update.test.ts` — V2 generational-cache validation, native package update, recovery, and focused coverage
+- `tui.tsx` — RPC status client, connecting/readiness and partition UI, idle refresh, and notification listener
+- `update.ts` — V2 generational-cache validation, native package update, and recovery
 - `utils.ts` — Generic result type and constructors
+
+### `packages/opencode-v2/test/`
+
+- `index.test.ts` — Plugin setup, command registration, update trigger, and context injection coverage
+- `commands.test.ts` — User-command preservation, ordered Helper execution, output framing, and prompt admission coverage
+- `preamble.test.ts` — Session immutability coverage
+- `tui.test.ts` — Status display, slot registration, toasts, and idle refresh coverage
+- `update.test.ts` — Cache validation, version comparison, update, and recovery coverage
 
 ### `packages/opencode-v2/scripts/`
