+++ .operator-shared/index/opencode.md
@@ -23,22 +23,25 @@ read_if: Working in packages/opencode or changing OpenCode preamble injection
 ## `packages/opencode/` Index
 
 - `README.md` — Published npm landing page with Helper-managed installation, OpenCode commands, session guidance, and documentation links
-- `tsconfig.check.json` — Package-scoped source, test, and build-script typecheck configuration
+- `tsconfig.check.json` — Package-scoped source and build-script typecheck configuration
 
 ### `packages/opencode/src/`
 
 - `index.ts` — OpenCode plugin entrypoint, detached Helper update trigger, hook wiring, fatal error handling, and deduplicated recovery notices
-- `index.test.ts` — Plugin startup coverage for detached, deduplicated Helper update checks
 - `tui.tsx` — Installation-aware home readiness indicator, sidebar partition status, per-project caching, and completed-turn refresh
-- `tui.test.ts` — TUI slot registration and completed-turn refresh coverage
 - `client.ts` — V1 plugin-client to V2 SDK client bridge
 - `commands.ts` — Non-overwriting inline setup and memory-repair command registration
-- `commands.test.ts` — Command sequencing, user-command preservation, configuration, and repair-instruction coverage
 - `preamble.ts` — Result-bearing preamble loading, rendering, and immutable normal or diagnostic per-session caching
-- `preamble.test.ts` — Normal and recovery-diagnostic session immutability coverage
 - `update.ts` — Nonblocking npm version check and direct stable `@latest` wrapper update with restart notification
-- `update.test.ts` — Wrapper validation, semantic-version comparison, update, and no-op coverage
 - `utils.ts` — Generic result type and success/failure constructors
+
+### `packages/opencode/test/`
+
+- `index.test.ts` — Plugin startup coverage for detached, deduplicated Helper update checks
+- `tui.test.ts` — TUI slot registration and completed-turn refresh coverage
+- `commands.test.ts` — Command sequencing, user-command preservation, configuration, and repair-instruction coverage
+- `preamble.test.ts` — Normal and recovery-diagnostic session immutability coverage
+- `update.test.ts` — Wrapper validation, semantic-version comparison, update, and no-op coverage
 
 ### `packages/opencode/scripts/`
